requirements.txt
cryptography>=41.0.0
pynacl>=1.5.0
requests>=2.31.0
flask>=3.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
click>=8.1.0
pytest>=7.4.0
